requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
django==5.0.3
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.1
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
sqlparse==0.4.4
    # via django
